+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxInfo.hs
@@ -149,7 +149,7 @@ import Cardano.Ledger.SafeHash (SafeHash, extractHash, hashAnnotated)
 import qualified Cardano.Ledger.Shelley.HardForks as HardForks
 import Cardano.Ledger.Shelley.TxCert
 import Cardano.Ledger.TxIn (TxId (..), TxIn (..))
-import Cardano.Ledger.UTxO (UTxO (..))
+import Cardano.Ledger.UTxO (EraUTxO (getScriptsProvided), ScriptsProvided (..), UTxO (..))
 import Cardano.Ledger.Val (Val (..))
 import Cardano.Slotting.EpochInfo (EpochInfo, epochInfoSlotToUTCTime)
 import Cardano.Slotting.Slot (EpochNo (..), SlotNo (..))
@@ -538,6 +538,12 @@ class ExtendedUTxO era where
     UTxO era ->
     Tx era ->
     Map.Map (ScriptHash (EraCrypto era)) (Script era)
+  default txscripts ::
+    EraUTxO era =>
+    UTxO era ->
+    Tx era ->
+    Map.Map (ScriptHash (EraCrypto era)) (Script era)
+  txscripts utxo = unScriptsProvided . getScriptsProvided utxo
 
   getAllowedSupplimentalDataHashes ::
     TxBody era ->
@@ -558,6 +564,7 @@ class ExtendedUTxO era where
   default getDatum :: AlonzoEraUTxO era => Tx era -> UTxO era -> ScriptPurpose era -> Maybe (Data era)
   getDatum tx utxo = getSpendingDatum utxo tx
 
+{-# DEPRECATED txscripts "In favor of `getScriptsProvided`" #-}
 {-# DEPRECATED getAllowedSupplimentalDataHashes "In favor of `getSupplementalDataHashes`" #-}
 {-# DEPRECATED getDatum "In favor of `getDatumForSpending`" #-}