+++ eras/conway/impl/src/Cardano/Ledger/Conway.hs
@@ -19,7 +19,6 @@ where
 import Cardano.Ledger.Alonzo (reapplyAlonzoTx)
 import Cardano.Ledger.Alonzo.Language (Language (..))
 import Cardano.Ledger.Alonzo.TxInfo (EraPlutusContext, ExtendedUTxO (..))
-import Cardano.Ledger.Babbage.Tx (babbageTxScripts)
 import Cardano.Ledger.Babbage.TxBody ()
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Era (ConwayEra)
@@ -70,4 +69,3 @@ instance
   ExtendedUTxO (ConwayEra c)
   where
   txInfo = conwayTxInfo
-  txscripts = babbageTxScripts